+++ src/toutdansunpackage/thing/personnage/Joueur.java
@@ -11,7 +11,7 @@ public class Joueur extends Personnage {
 		super(startPosition, startDirection);
 		vie = 100;
 		armure = 50;
-		v=0.05;
+		v=(float) 0.05;
 	}		
 	
 	public Joueur() {}
